requirements.txt
pandas
matplotlib
seaborn
plotly
wordcloud
PyGithub
GitPython
jinja2
weasyprint
nltk
python-dotenv
pyyaml
